+++ src/Command/RunConsumerCommand.php
@@ -14,6 +14,8 @@ declare(strict_types = 1);
 namespace FiveLab\Component\Amqp\Command;
 
 use FiveLab\Component\Amqp\Consumer\ConsumerInterface;
+use FiveLab\Component\Amqp\Consumer\Middleware\StopAfterNExecutesMiddleware;
+use FiveLab\Component\Amqp\Consumer\MiddlewareAwareInterface;
 use FiveLab\Component\Amqp\Consumer\Registry\ConsumerRegistryInterface;
 use FiveLab\Component\Amqp\Exception\ConsumerTimeoutExceedException;
 use Symfony\Component\Console\Command\Command;
@@ -56,7 +58,8 @@ class RunConsumerCommand extends Command
             ->setDescription('Run consumer.')
             ->addArgument('key', InputArgument::REQUIRED, 'The key of consumer.')
             ->addOption('read-timeout', null, InputOption::VALUE_REQUIRED, 'Set the read timeout for RabbitMQ.')
-            ->addOption('loop', null, InputOption::VALUE_NONE, 'Loop consume (used only with read-timeout).');
+            ->addOption('loop', null, InputOption::VALUE_NONE, 'Loop consume (used only with read-timeout).')
+            ->addOption('messages', null, InputOption::VALUE_REQUIRED, 'After process number of messages process be normal exits.');
     }
 
     /**
@@ -69,6 +72,18 @@ class RunConsumerCommand extends Command
         // Verify input parameters
         if ($input->getOption('loop') && !$input->getOption('read-timeout')) {
             throw new \InvalidArgumentException('The "read-timeout" is required for loop consume.');
+        }
+
+        if ($input->getOption('messages')) {
+            if (!$consumer instanceof MiddlewareAwareInterface) {
+                throw new \InvalidArgumentException(\sprintf(
+                    'For set number of messages customer must implement "%s", but "%s" given.',
+                    MiddlewareAwareInterface::class,
+                    \get_class($consumer)
+                ));
+            }
+
+            $consumer->pushMiddleware(new StopAfterNExecutesMiddleware((int) $input->getOption('messages')));
         }
 
         if ($input->getOption('read-timeout')) {