+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 CHANGELOG
 =========
 
+v1.1.2
+--------
+
+* Add option `--messages` to run consumer command.
+
+v1.1.1
+--------
+
+* Add logging consumer.
+
+v1.1.0
+--------
+
+* Add method for get count messages from queue.
+* Add delay system (allow publish messages with delay).
+* Add command for list possible consumers.
+
 v1.0.0
 ------